Restructure project a little
