+++ src/features/editor/components/Sidebar.module.css
