+++ src/features/simulator/Simulator.module.css
