+++ src/features/simulator/components/Edge.module.css
