+++ src/features/simulator/components/Vertex.tsx
